multiple files: add start/stop flags to mspike_extract script
Ignore-this: a429675cb2249450c0fa480748f2e212

darcs-hash:20100616182517-21d25-f8a49ef15414c1376182324dba73298aa4b5c98a.gz

+++ mspikes/groupevents.py
@@ -204,9 +204,9 @@ def main():
         elif o == '--stimulus':
             options['stimuli'] = tuple(x.strip() for x in a.split(','))
         elif o == '--start':
-            options['start'] = int(a)
+            options['start'] = float(a)
         elif o == '--stop':
-            options['stop'] = int(a)
+            options['stop'] = float(a)
         elif o == '-b':
             options['basename'] = a
